Manually check BS samples match master list
Added script to line up sample names from BS sequencing with the top
genotype matches from SNPmatch, as well as with what genotypes they
ought to be from the field master list. This saves a CSV file that
I manually checked.

Plate 145 was done at a time when Viktoria could not process a 96-
well plate at once, so did 24 wells four times. I could not find
a clear description of how to line those up with the samples
programatically, so I did it by hand.

 On branch process_reads
 Changes to be committed:
	new file:   003.scripts/006.get_plate_positions.py
	new file:   004.output/manually_check_snpmatch_results.csv