Script to get genome-wide methylation
SLURM script to calculate weighted-mean methylation levels in CG,
CHG and CHH sequence contexts for a folder of allc files generated
by the methylpy pipeline.

This loops over each allc file, calculates numbers, and sends them
to an output summary file. This would be faster as a job array, but
I wanted to be sure that the output file was created and updated
within a single script.

 On branch process_allc
 Changes to be committed:
	new file:   003.scripts/007.mean_mC_jobarray.sh